Extract Msf::DBManager::Import::MBSA
MSP-11124
